detect wallets on run
